+++ pkg/model/logger_config.go
@@ -1,10 +1,5 @@
 package model
 
-import (
-	"errors"
-	"fmt"
-)
-
 // LoggerConfig represents the backup service logger configuration.
 // @Description LoggerConfig represents the backup service logger configuration.
 //
@@ -47,31 +42,6 @@ func (l *LoggerConfig) GetStdoutWriterOrDefault() bool {
 	return *defaultConfig.logger.StdoutWriter
 }
 
-// Compare compares two LoggerConfig structs and returns an error if they differ.
-func (l *LoggerConfig) Compare(other *LoggerConfig) error {
-	if l == nil && other == nil {
-		return nil
-	}
-	if l == nil {
-		return errors.New("logger added")
-	}
-	if other == nil {
-		return errors.New("logger removed")
-	}
-
-	var err = errors.Join(
-		comparePointers("Level", l.Level, other.Level),
-		comparePointers("Format", l.Format, other.Format),
-		comparePointers("StdoutWriter", l.StdoutWriter, other.StdoutWriter),
-	)
-
-	if e := l.FileWriter.Compare(other.FileWriter); e != nil {
-		err = errors.Join(err, fmt.Errorf("FileWriter changes: %w", e))
-	}
-
-	return err
-}
-
 // FileLoggerConfig represents the configuration for the file logger writer.
 // @Description FileLoggerConfig represents the configuration for the file logger writer.
 type FileLoggerConfig struct {
@@ -89,25 +59,4 @@ type FileLoggerConfig struct {
 	// Compress determines if the rotated log files should be compressed
 	// using gzip. The default is not to perform compression.
 	Compress bool
-}
-
-// Compare FileLoggerConfig with another and return detailed errors.
-func (f *FileLoggerConfig) Compare(other *FileLoggerConfig) error {
-	if f == nil && other == nil {
-		return nil
-	}
-	if f == nil {
-		return fmt.Errorf("FileLogger added")
-	}
-	if other == nil {
-		return fmt.Errorf("FileLogger removed")
-	}
-
-	return errors.Join(
-		compareValues("Filename", f.Filename, other.Filename),
-		compareValues("MaxSize", f.MaxSize, other.MaxSize),
-		compareValues("MaxAge", f.MaxAge, other.MaxAge),
-		compareValues("MaxBackups", f.MaxBackups, other.MaxBackups),
-		compareValues("Compress", f.Compress, other.Compress),
-	)
 }
